requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2
requests>=2.30
python-dotenv>=1.0
APScheduler>=3.10
tqdm>=4.65
pyarrow>=14.0
google-cloud-storage>=2.14
